frame_t : Change type of frame index from double to int.

+++ headers/frame_t.h
@@ -18,7 +18,7 @@ namespace tmd {
 
     typedef struct {
         cv::Mat original_frame;         // Original frame taken from the video.
-        double frame_index;             // Index of the frame in the video.
+        int frame_index;             // Index of the frame in the video.
         cv::Mat mask_frame;             // Frame after applying background substraction.
         int camera_index;     // Index of the camera which took the frame.
         std::vector<tmd::player_t*> players; // Players on the frame.
